requirements.txt
numpy
pandas
pyarrow
openpyxl
requests

# LlamaIndex core + embeddings + Qdrant vector store
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-index-postprocessor-flag-embedding-reranker
sentence-transformers
qdrant-client
groq

# Reranker (optional)
FlagEmbedding

# PDF generation
reportlab

# File loaders
pymupdf

# Telegram bot
python-telegram-bot

# Instagram
instagrapi==2.2.1

# OpenAI-compatible client for Groq
# openai (optional if using OpenAI-compatible libs elsewhere)

# Runtime helpers
python-dotenv
gradio